transform return: Removed an optimization that broke some table literals
This optimization would normally remove an unnecessary variable, such as
the one in the following generated PIR:

set $P18, $P17[k_Foo]
new $P19, "LuaNil"
.tailcall $P18()

In the above snippet, $P19 is not used, so it can be removed without
harm. However, there are instances where this variable is used. Consider
the following code:

function Foo()
	return 42
end

return { Foo() }

In this case, the following PIR should be generated:

new $P21, "LuaNil"
($P21 :slurpy) = $P20()
.tailcall tconstruct($P18, $P19, $P21 :flat)

In this case, the register that initially contains a nil value is
actually used, so it cannot be optimized away.